Move offsetLine to query_utils in preparation of unit test
